CPM-799: Add tests for identifier generator context change
