+++ components/identifier-generator/front/src/feature/context/IdentifierGeneratorContext.ts
@@ -3,14 +3,13 @@ import {createContext} from 'react';
 type IdentifierGeneratorContextType = {
   unsavedChanges: {
     hasUnsavedChanges: boolean;
-    setHasUnsavedChanges: (newValue: boolean) => void;
+    setHasUnsavedChanges?: (newValue: boolean) => void;
   };
 };
 
 const defaultValue = {
   unsavedChanges: {
     hasUnsavedChanges: false,
-    setHasUnsavedChanges: () => null,
   },
 };
