+++ components/identifier-generator/front/src/feature/pages/__tests__/CreateGeneratorPage.test.tsx
@@ -5,19 +5,9 @@ import {IdentifierGenerator} from '../../models';
 import {Router} from 'react-router-dom';
 import {act, fireEvent, waitFor} from '@testing-library/react';
 import {createMemoryHistory} from 'history';
+import {initialGenerator} from '../../tests/fixtures/initialGenerator';
 
 jest.mock('../CreateOrEditGeneratorPage');
-
-const initialGenerator: IdentifierGenerator = {
-  code: 'initialCode',
-  labels: {
-    en_US: 'Initial Label',
-  },
-  conditions: [],
-  structure: [],
-  delimiter: null,
-  target: 'sku',
-};
 
 describe('CreateGeneratorPage', () => {
   it('should create a generator', async () => {
