+++ components/identifier-generator/front/src/feature/pages/__tests__/CreateOrEditGeneratorPage.test.tsx
@@ -1,25 +1,14 @@
 import React from 'react';
 import {fireEvent, render, screen} from '../../tests/test-utils';
 import {CreateOrEditGeneratorPage} from '../';
-import {IdentifierGenerator, PROPERTY_NAMES} from '../../models';
 import {createMemoryHistory} from 'history';
 import {Router} from 'react-router';
 import {IdentifierGeneratorContext} from '../../context/IdentifierGeneratorContext';
+import {initialGenerator} from '../../tests/fixtures/initialGenerator';
 
 jest.mock('../DeleteGeneratorModal');
 jest.mock('../../tabs/GeneralPropertiesTab');
 jest.mock('../../tabs/StructureTab');
-
-const initialGenerator: IdentifierGenerator = {
-  code: 'initialCode',
-  labels: {
-    en_US: 'Initial Label',
-  },
-  conditions: [],
-  structure: [{type: PROPERTY_NAMES.FREE_TEXT, string: 'AKN'}],
-  delimiter: null,
-  target: 'sku',
-};
 
 describe('CreateOrEditGeneratorPage', () => {
   it('should switch tabs', () => {
@@ -104,24 +93,6 @@ describe('CreateOrEditGeneratorPage', () => {
   });
 
   it('should update a generate on structure change', () => {
-    render(
-      <CreateOrEditGeneratorPage
-        isMainButtonDisabled={false}
-        initialGenerator={initialGenerator}
-        validationErrors={[]}
-        mainButtonCallback={jest.fn()}
-        isNew={false}
-      />
-    );
-
-    fireEvent.click(screen.getByText('pim_identifier_generator.tabs.identifier_structure'));
-    expect(screen.getByText('StructureTabMock')).toBeInTheDocument();
-    expect(screen.getByText('[{"type":"free_text","string":"AKN"}]')).toBeInTheDocument();
-    fireEvent.click(screen.getByText('Update Free Text'));
-    expect(screen.getByText('[{"type":"free_text","string":"Updated string"}]')).toBeInTheDocument();
-  });
-
-  it('should update identifier generator context on change', () => {
     const mockSetHasUnsavedChanges = jest.fn();
     const identifierGeneratorDependencies = {
       unsavedChanges: {
@@ -130,23 +101,27 @@ describe('CreateOrEditGeneratorPage', () => {
       },
     };
 
-    render(<>
+    render(
       <IdentifierGeneratorContext.Provider value={identifierGeneratorDependencies}>
-          <CreateOrEditGeneratorPage
-            isMainButtonDisabled={false}
-            initialGenerator={initialGenerator}
-            mainButtonCallback={jest.fn()}
-            validationErrors={[]}
-            isNew={false}
-          />
+        <CreateOrEditGeneratorPage
+          isMainButtonDisabled={false}
+          initialGenerator={initialGenerator}
+          validationErrors={[]}
+          mainButtonCallback={jest.fn()}
+          isNew={false}
+        />
       </IdentifierGeneratorContext.Provider>
-    </>);
+    );
 
     fireEvent.click(screen.getByText('pim_identifier_generator.tabs.identifier_structure'));
     expect(screen.getByText('StructureTabMock')).toBeInTheDocument();
     expect(screen.getByText('[{"type":"free_text","string":"AKN"}]')).toBeInTheDocument();
     fireEvent.click(screen.getByText('Update Free Text'));
+    expect(screen.getByText('[{"type":"free_text","string":"Updated string"}]')).toBeInTheDocument();
 
-    expect(mockSetHasUnsavedChanges).toHaveBeenCalled();
+    expect(mockSetHasUnsavedChanges).toHaveBeenCalledWith(true);
+
+    fireEvent.click(screen.getByText('Revert Free Text'));
+    expect(mockSetHasUnsavedChanges).toHaveBeenCalledWith(false);
   });
 });
