+++ components/identifier-generator/front/src/feature/pages/__tests__/EditGeneratorPage.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import {mockResponse, render, screen, waitFor} from '../../tests/test-utils';
-import {IdentifierGenerator, PROPERTY_NAMES} from '../../models';
 import {EditGeneratorPage} from '../';
 import userEvent from '@testing-library/user-event';
 import {NotificationLevel} from '@akeneo-pim-community/shared';
+import {initialGenerator} from '../../tests/fixtures/initialGenerator';
 
 const mockNotify = jest.fn();
 
@@ -19,22 +19,6 @@ jest.mock('@akeneo-pim-community/shared', () => ({
     return mockNotify;
   },
 }));
-
-const initialGenerator: IdentifierGenerator = {
-  code: 'initialCode',
-  labels: {
-    en_US: 'Initial Label',
-  },
-  conditions: [],
-  structure: [
-    {
-      type: PROPERTY_NAMES.FREE_TEXT,
-      string: 'AKN',
-    },
-  ],
-  delimiter: null,
-  target: 'sku',
-};
 
 const defaultUiLocales = [
   {
