+++ components/identifier-generator/front/src/feature/tabs/__mocks__/StructureTab.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import {FreeText, Structure} from '../../models';
+import {PROPERTY_NAMES, Structure} from '../../models';
+import {initialGenerator} from '../../tests/fixtures/initialGenerator';
 
 type StructureTabProps = {
   onStructureChange: (structure: Structure) => void;
@@ -8,15 +9,19 @@ type StructureTabProps = {
 
 const StructureTab: React.FC<StructureTabProps> = ({onStructureChange, initialStructure}) => {
   const updateFreeText = () => {
-    (initialStructure[0] as FreeText).string = 'Updated string';
-    onStructureChange(initialStructure);
+    const updatedStructure = [...initialStructure];
+    updatedStructure[0] = {type: PROPERTY_NAMES.FREE_TEXT, string: 'Updated string'};
+    onStructureChange(updatedStructure);
   };
+
+  const revertFreeText = () => onStructureChange(initialGenerator.structure);
 
   return (
     <>
       StructureTabMock
       <div>{JSON.stringify(initialStructure)}</div>
       <button onClick={updateFreeText}>Update Free Text</button>
+      <button onClick={revertFreeText}>Revert Free Text</button>
     </>
   );
 };
