+++ components/identifier-generator/front/src/feature/tests/fixtures/initialGenerator.ts
@@ -1,0 +1,14 @@
+import {IdentifierGenerator, PROPERTY_NAMES} from '../../models';
+
+const initialGenerator: IdentifierGenerator = {
+  code: 'initialCode',
+  labels: {
+    en_US: 'Initial Label',
+  },
+  conditions: [],
+  structure: [{type: PROPERTY_NAMES.FREE_TEXT, string: 'AKN'}],
+  delimiter: null,
+  target: 'sku',
+};
+
+export {initialGenerator};
